[AGREEMENT] define the purpose and accountability of the devops team
The main change is that devops are held accountable for keeping the
credentials they use safe.

This was previously what "Secret Keepers" were accountable for. But
the reality is that devops need root access to everything in order to
perform their duties and keep things running / functional from a
technical point of view.

The "Secret Keepers", "Admins" and "Chatrooms / matrix" teams are
merged together because keeping them separated just adds
administrative overhead at this point.

* https://matrix.to/#/#forgejo-releases:a-0.me
  * https://matrix.to/#/@ph:a-0.me

Is the only exception because it is a matrix room managed by a single
individual. It is linked from the Forgejo Space but is a non essential
service and if becomes unmaintained in the future, it could be
disconnected. The person maintaining this room clearly stated in the
past that they are not willing to be subject to the administrative
overhead of being formally part of a Forgejo team.