TEAMS: document the current contributors membership and the process
The membership of the contributors team has been more relaxed than for
the security team or the release team. It does not involve permissions
that require the same level of scrutiny.

The current list of contributors is updated for information, together
with the permissions they have on the associated repositories.

Since the membership of this team does not go through the decision
making process, it should also be clarified that members can be asked
to go through the decision process to become a member. The stakes are
very low and it is unlikely that such a scenario will ever happen. But
in case it does, it is better to have it in writing.