2024 team membership
I declared my intention to spend 4 hours a week on Forgejo in
2024[0]. My focus will be on F3[1] and helping with the Forgejo's
driver implementation.

I also intend to participate in diverse Forgejo related activities
but none of them require me to be part of any team. Since the renewal
of my membership is due in the following weeks, this is as good a time
as any to update my team membership all at once.

[0] https://codeberg.org/forgejo/sustainability/pulls/35
[1] https://f3.forgefriends.org