link moderation reports from the README
The moderation reports could then be closed as there will be a link to
find them from the README. It is good to have a way to list them all
but it clutters the display of governance issues to have all
moderation reports open forever.

The moderation reports older than three months were closed.