teams: earl-warren stepping down from moderation
It has been a year since I started and it was a time of hardship. To be effective in this role I need a peace of mind that is inevitably eroded by being exposed to behavior that goes against the code of conduct or the law. I will apply again in 2027.