Support remote pages for chart and tabular data pages
Note that this does not yet include support for change propagation
from the central repository to the remote wikis.

Data fetches are done via the action API on the JsonConfig backend,
so in principle this will support fetching definitions and data from
Commons on third-party sites as well, however it may not compose well
with local data sets.

If we want to support using local Data: pages in a local Data: ns
that transparently falls back to remote Commons Data: pages like how
images work, we may need to make some changes to JsonConfig. It
seems to prefer that they be separate patterns or namespaces rather
than composing as layers.

Readme includes an example config that seems to be working locally
with two sites running on a shared codebase.

Bug: T371924
Change-Id: I34ccc278337e9ae530c0517cac934cd397004507