Fix E_NOTICE when parsing {{#chart:}} without arguments
This happened because we were using strpos() incorrectly: when it
doesn't find what it's looking for, it returns false instead of -1, and
false is also considered to be >= 0. Instead, use the more modern (and
clearer) str_contains() function.

Also use trim() so that spaces around the |s and =s don't break things.

Change-Id: I2a4854d07f88a54a995d430a1fb46427d9ba2732