XENS: BACKPORT: RenameUser: Add derived option to skip updating shared tables.
When renaming a user, RenameuserSQL will be (since next change) called
on each wiki within the family, so that all local (non-shared and not in
a mapped virtual domain) tables can be updated.
This is necessary to avoid updating tables that are set to be shared
with virtual domains for multiple times.

Bug: T104830
Change-Id: Ic4c4838ab6663594b35c0286dd0d5e168a2f6432