ADR for usage tracking table
Local-only setups will continue to use templatelinks, while remote
systems will use the shared table proposed in this ADR. An API mechanism
will be added to propagate cache invalidations from the store wiki
to remote wikis that have recorded usages.

Note most of the actual work will live in JsonConfig, so we'll
want to copy the final ADR in there as well.

Bug: T370378
Change-Id: I8a892275a2f7484b5c8898b04ab04160a1868852