language: Use fallback chain to create NumberFormatter
php8.4 validates the locale and rejects invalid one.
Just use the fallback chain to find an usable language, mostly en.
This is a no-op for php < 8.4, but can impact performance for some
languages due to more attempts to create a NumberFormatter.
To cache the NumberFormatter the cache key must include some of the
called setter on that object

Also C is rejected, so resolve it directly with locale_get_default()

Bug: T376711
Change-Id: I686f77baea33ea2852c546e30e9cc08618d44321